Create Prime Numbers.cbp
In the set of Natural numbers i.e positive integers , a prime number is either divisible by itself or by one. In the program the input is divided by i, if i divides n strictly 2 times then the number is prime. Otherwise, it is a composite number having at least 3 divisors.